+++ src/apps/seelenweg/modules/bar/app.ts
@@ -1,17 +1,10 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { SeelenWegMode, SeelenWegSide, SeelenWegState } from '../../../settings/modules/seelenweg/domain';
+import { SeelenWegSlice as SeelenWegSettingsSlice } from '../../../settings/modules/seelenweg/app';
 
-const initialState: SeelenWegState = {
-  enabled: true,
-  mode: SeelenWegMode.MIN_CONTENT,
-  position: SeelenWegSide.BOTTOM,
-  size: 40,
-  zoomSize: 70,
-  margin: 8,
-  padding: 8,
-  spaceBetweenItems: 8,
-};
+import { SeelenWegState } from '../../../settings/modules/seelenweg/domain';
+
+const initialState: SeelenWegState = SeelenWegSettingsSlice.getInitialState();
 
 export const SeelenWegSlice = createSlice({
   name: 'seelenweg',